+++ src/books/service/books.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
-import { BookModel } from 'src/models/entity/book.entity';
+import { BookModel } from '../../models/entity/book.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 
 @Injectable()
@@ -18,16 +18,16 @@ export class BooksService {
     return await this.booksRepository.find();
   }
 
-  async update(id: string, newBook: BookModel): Promise<BookModel> {
+  async update(id: string, newBook: BookModel): Promise<BookModel | undefined> {
     const bookToUpdate = await this.booksRepository.findOne(id);
-    if (!bookToUpdate) return;
-    const updatedBook = {...bookToUpdate, ...newBook};
-    return await this.booksRepository.save(updatedBook);
+    if (bookToUpdate) {
+      const updatedBook = {...bookToUpdate, ...newBook};
+      return await this.booksRepository.save(updatedBook);
+    }
   }
 
-  async delete(id: string): Promise<BookModel> {
+  async delete(id: string): Promise<BookModel | undefined> {
     const bookToRemove = await this.booksRepository.findOne(id);
-    if (!bookToRemove) return;
-    return await this.booksRepository.remove(bookToRemove);
+    if (bookToRemove) return await this.booksRepository.remove(bookToRemove);
   }
 }
